chore(US-004): auto-commit after rectification session
