+++ src/operations/call.ts
@@ -410,6 +410,7 @@ export async function callOp<I, O, C>(ctx: CallContext, op: Operation<I, O, C>, 
 
   const rawOutput = outcome.result.output;
   const totalCost = outcome.result.estimatedCostUsd ?? 0;
+  ctx.onCostAccumulated?.(totalCost);
 
   if (!rawOutput) {
     throw new NaxError(`callOp[${op.name}]: agent returned no output`, "CALL_OP_NO_OUTPUT", {
